queue only the most recent pending message
In `poll_next`, a response message may be queued if the sink is not
ready. However, as only close (after which no further messages may be
sent) and pong messages (which the endpoint MAY elect to send only in
response to the most recently processed ping message), only the most
recent message needs to be queued.

This makes the code much easier to reason about. If desired, responding
to all ping messages could be configurable.