do not echo close message's reason
The RFC has this to say for this branch:
"If an endpoint receives a Close frame and did not previously send a
Close frame, the endpoint MUST send a Close frame in response. (When
sending a Close frame in response, the endpoint typically echos the
status code it received.)" Echoing the reason is typically not done and
serves no benefit as its already known to the peer.